Iterative chunking implemented
Could probably find a couple of optimizations before attempting threading

+++ src/cubemarch.h
@@ -36,7 +36,7 @@ void marchCube(
     int x, int y, int z,
     std::vector<float>& noiseValues, int size,
     std::vector<Vector3>& vertices,
-    std::vector<size_t>& indices,
+    std::vector<int>& indices,
     std::vector<int>* edgeCacheLocal,
     std::vector<int>* edgeCacheX,
     std::vector<int>* edgeCacheY,
